requirements.txt
cachetools==5.4.0
colorama==0.4.6
google-auth==2.32.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pytz==2024.1
requests-oauthlib==2.0.0
rsa==4.9
